Testing: Move TestRucioServer under tests; fix #7255
